MongoSource uses Observable to Publisher adapter explicitly instead of implicitly
