+++ mongodb/src/main/scala/akka/stream/alpakka/mongodb/scaladsl/MongoSource.scala
@@ -1,13 +1,13 @@
 package akka.stream.alpakka.mongodb.scaladsl
 
-import akka.stream.alpakka.mongodb.Implicits._
+import akka.stream.alpakka.mongodb.ObservableToPublisher
 import akka.stream.scaladsl.Source
 import org.mongodb.scala.{Document, Observable}
 
 object MongoSource {
 
-  def apply(query: Observable[Document]) ={
-    Source.fromPublisher(query)
+  def apply(query: Observable[Document]) = {
+    Source.fromPublisher(ObservableToPublisher(query))
   }
 
 }
